requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
Brotli 
certifi 
cffi==1.17.0
charset-normalizer 
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
datasets==3.0.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
executing==2.1.0
filelock
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna 
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2 
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.5.0
keyboard==0.13.5
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
langgraph==0.2.28
langgraph-checkpoint==1.0.12
langsmith==0.1.120
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-fft
mkl-random 
mkl-service==2.4.0
ml-dtypes==0.4.1
more-itertools==10.4.0
mpmath 
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx
numba==0.60.0
numpy 
openai-whisper==20231117
opt-einsum==3.3.0
optimum==1.22.0
optree==0.12.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
peft==0.12.0
pillow 
platformdirs==4.3.3
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
pyreadline3==3.5.2
PySocks 
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML
pyzmq==26.2.0
regex==2024.7.24
requests
rich==13.8.1
rouge==1.0.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
sentencepiece==0.2.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.34
stack-data==0.6.3
sympy 
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions 
tzdata==2024.1
urllib3 
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.43.0
widgetsnbextension==4.0.13
win-inet-pton 
wrapt==1.16.0
xxhash==3.5.0
yarl==1.11.1
